getByStatus terminado y regexp para validar email

+++ src/main/java/com/biblioteca/big/controller/BookController.java
@@ -1,13 +1,16 @@
 package com.biblioteca.big.controller;
 
+import com.biblioteca.big.exception.BookNotFoundException;
 import com.biblioteca.big.model.Book;
 import com.biblioteca.big.repository.BookRepository;
 import com.biblioteca.big.service.BookService;
+import com.sun.source.tree.LambdaExpressionTree;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -27,23 +30,41 @@ public class BookController {
     }
 
     @GetMapping("/{id}")
-    public Book getBookById(@PathVariable Long id){
+    public Book getBookById(@PathVariable Long id) throws BookNotFoundException {
         Optional<Book> book = bookRepository.findById(id);
+
+        if(!book.isPresent()){
+            throw new BookNotFoundException("No se encontró el ID: " + id);
+        }
+
         return book.get();
-        //TODO: MANEJAR LA EXCEPCION SI NO EXISTE EL ID
     }
 
-    /*
-    @GetMapping("/{status}")
-    public List<Book> getBooksByStatus(){
-        return bookService.getByStatus(Sort.by("title").ascending());
+
+    @GetMapping("/status/{status}")
+    public List<Book> getBooksByStatus(@PathVariable("status") String status){
+        List<Book> books = new ArrayList<>();
+
+        if(status.equals("DISPONIBLE")){
+            books = bookRepository.findByStatus("DISPONIBLE", Sort.by("title").ascending());
+            return books;
+        }
+        else if (status.equals("RESERVADO")){
+            books = bookRepository.findByStatus("RESERVADO", Sort.by("title").ascending());
+            return books;
+        }
+        return books;
     }
-    */
 
     @DeleteMapping("/{id}")
-    public void deleteBookById(@PathVariable Long id){
+    public void deleteBookById(@PathVariable Long id) throws BookNotFoundException{
+        Optional<Book> book = bookRepository.findById(id);
+
+        if(!book.isPresent()){
+            throw new BookNotFoundException("No se encontró el ID: " + id);
+        }
+
         bookRepository.deleteById(id);
-        //TODO: MANEJAR LA EXCEPCION SI NO EXISTE EL ID
     }
 
     @PutMapping("/{id}")
